set city

+++ lib/models/meteo_forecast.dart
@@ -62,7 +62,7 @@ class Main {
     temp = json['temp'];
     feelsLike = json['feels_like'].toDouble();
     tempMin = json['temp_min'];
-    tempMax = json['temp_max'];
+    tempMax = json['temp_max'].toDouble();
     pressure = json['pressure'];
     seaLevel = json['sea_level'];
     grndLevel = json['grnd_level'];
@@ -105,7 +105,7 @@ class Wind {
   Wind({this.speed, this.deg, this.gust});
 
   Wind.fromJson(Map<String, dynamic> json) {
-    speed = json['speed'];
+    speed = json['speed'].toDouble();
     deg = json['deg'];
     gust = json['gust'];
   }
